+++ packages/cohort/src/model/mod.rs
@@ -1,4 +1,5 @@
 pub mod bank_account;
 pub(crate) mod json;
+pub mod requests;
 pub mod snapshot;
 pub mod talos_state;
