+++ packages/cohort/src/state/mod.rs
@@ -1,4 +1,3 @@
-pub mod model;
 pub mod postgres;
 #[cfg(test)]
 pub mod test_utils;
