+++ packages/cohort/src/tx_batch_executor.rs
@@ -2,9 +2,9 @@ use std::str::FromStr;
 
 use crate::actions::account_update::AccountUpdate;
 use crate::actions::transfer::Transfer;
+use crate::model::requests::{AccountUpdateRequest, BusinessActionType, TransferRequest};
 use crate::replicator::core::StatemapItem;
 use crate::snapshot_api::SnapshotApi;
-use crate::state::model::{AccountUpdateRequest, BusinessActionType, TransferRequest};
 use crate::state::postgres::database::{Action, Database};
 use deadpool_postgres::GenericClient;
 use futures::future::BoxFuture;
